refactored test classes
